refactor explore
