+++ components/explore/ExplorePage.jsx
@@ -1,7 +1,6 @@
 import { motion } from "framer-motion";
 import Link from "next/link";
 import fadeIn from "../../lib/framer/animations/fadeIn";
-import styles from "../../styles";
 
 const ToolsCard = ({
   id,
@@ -15,9 +14,9 @@ const ToolsCard = ({
 }) => (
   <motion.div
     variants={fadeIn("right", "spring", index * 0.5, 0.75)}
-    className={`relative ${
-      active === id ? "lg:flex-[3.5] flex-[10]" : "lg:flex-[0.5] flex-[2]"
-    } flex items-center justify-center min-w-[170px] h-[700px] transition-[flex] duration-[0.7s] ease-out-flex cursor-pointer`}
+    className={`relative min-w-[170px] h-[700px] flexCenter ${
+      active === id ? "flex-[10] lg:flex-[3.5]" : "flex-[2] lg:flex-[0.5]"
+    } transition-[flex] duration-[0.7s] ease-out-flex cursor-pointer`}
     onClick={() => handleClick(id)}
   >
     <img
@@ -26,32 +25,30 @@ const ToolsCard = ({
       className="absolute w-full h-full object-cover rounded-[24px]"
     />
     {active !== id ? (
-      <h3 className="font-semibold sm:text-[26px] text-[18px] text-white absolute z-0 lg:bottom-20 lg:rotate-[-90deg] lg:origin-[0,0]">
+      <h2 className="absolute z-0 lg:bottom-20 text-[18px] sm:text-[26px] font-semibold lg:rotate-[-90deg] lg:origin-[0,0] text-white">
         {title}
-      </h3>
+      </h2>
     ) : (
       <Link
         href={`explore/${query}`}
-        className="absolute bottom-0 p-8 flex justify-start w-full flex-col bg-[rgba(0,0,0,0.5)] rounded-b-[24px]"
+        className="absolute bottom-0 w-full p-8 flex flex-col justify-start rounded-b-[24px] bg-[rgba(0,0,0,0.5)]"
       >
-        <div
-          className={`${styles.flexCenter} w-[60px] h-[60px] rounded-[24px] glassmorphism mb-[16px]`}
-        >
+        <div className="w-[60px] h-[60px] flexCenter rounded-[24px] glassmorphism mb-[16px]">
           <img
-            src="/headset.svg"
+            src="/icons/headset.svg"
             alt="headset"
             className="w-1/2 h-1/2 object-contain"
           />
         </div>
-        <p className="font-normal text-[16px] leading-[20.16px] text-white uppercase">
+        <h3 className="text-[16px] leading-[20.16px] font-normal uppercase text-white">
           Cooming soon.
-        </p>
+        </h3>
         <h2 className="mt-[24px] font-semibold sm:text-[32px] text-[24px] text-white">
           {title}
         </h2>
-        <h2 className="font-normal text-[16px] leading-[20.16px] text-white uppercase">
+        <p className="font-normal text-[16px] leading-[20.16px] text-white uppercase">
           {desc}
-        </h2>
+        </p>
       </Link>
     )}
   </motion.div>
