+++ pages/explore/[projects].js
@@ -1,8 +1,8 @@
 import ErrorPage from "next/error";
 import { useRouter } from "next/router";
 import Layout from "../../components/Layout";
-import HeroExploreProjects from "../../components/explore/HeroExploreProjects";
-import MainExploreProjects from "../../components/explore/MainExploreProjects";
+import HeroExploreProjects from "../../partials/Explore/HeroExploreProjects";
+import MainExploreProjects from "../../partials/Explore/MainExplore/MainExploreProject";
 
 export default function Page() {
   const allowedSlugs = ["onramps", "bridge", "tools", "social", "defi", "nfts"];
