+++ partials/Explore/MainExplore/data/onramps.js
@@ -1,0 +1,18 @@
+const onramps = [
+  {
+    id: 1,
+    name: "Obiter Finance",
+    website: "https://www.orbiter.finance/",
+    desc: "Orbiter Finance is a decentralized cross-rollup bridge for transferring the Ethereum native assets, which is the infrastructure of Layer 2, it offers low cost and almost instant transfers.",
+    image: "/Bridge/obiter.png",
+  },
+  {
+    id: 2,
+    name: "Stargate Finance",
+    website: "https://www.orbiter.finance/",
+    desc: "Orbiter Finance is a decentralized cross-rollup bridge for transferring the Ethereum native assets, which is the infrastructure of Layer 2, it offers low cost and almost instant transfers.",
+    image: "/Bridge/obiter.png",
+  },
+];
+
+export default onramps;
