--- sections/explore/Hero.jsx
@@ -1,32 +1,0 @@
-"use client";
-
-import { motion } from "framer-motion";
-import staggerContainer from "../../lib/framer/staggerContainer";
-import textVariant from "../../lib/framer/text/textVariant";
-import styles from "../../styles";
-
-const Hero = () => (
-  <section className={`${styles.yPaddings} sm:pl-16 pl-6 `}>
-    <motion.div
-      variants={staggerContainer}
-      initial="hidden"
-      whileInView="show"
-      viewport={{ once: false, amount: 0.25 }}
-      className={`${styles.innerWidth} mx-auto flex flex-col`}
-    >
-      <div className="flex justify-center items-center flex-col relative z-10">
-        <motion.h1 variants={textVariant(1.1)} className={styles.heroHeading}>
-          EXPLORE
-        </motion.h1>
-        <motion.div
-          variants={textVariant(1.2)}
-          className="flex flex-row justify-center items-center"
-        >
-          <h1 className={styles.heroHeading}> Interchain</h1>
-        </motion.div>
-      </div>
-    </motion.div>
-  </section>
-);
-
-export default Hero;
